color refactor and copy change
