+++ packages/ui/src/pages/Home/Home.tsx
@@ -1,19 +1,14 @@
 import { PropsWithChildren, useCallback, useEffect, useState } from 'react'
 import { Box, Grid } from '@mui/material'
 import { useMultiProxy } from '../../contexts/MultiProxyContext'
-import { useNavigate, useSearchParams } from 'react-router-dom'
-import { Button } from '../../components/library'
+import { useSearchParams } from 'react-router-dom'
 import SuccessCreation from '../../components/SuccessCreation'
 import NewMulisigAlert from '../../components/NewMulisigAlert'
 import { styled } from '@mui/material/styles'
-import { Center } from '../../components/layout/Center'
-import { useAccounts } from '../../contexts/AccountsContext'
-import { useWatchedAddresses } from '../../contexts/WatchedAddressesContext'
 import HeaderView from './HeaderView'
 import MultisigView from './MultisigView'
 import TransactionList from '../../components/Transactions/TransactionList'
-import { ConnectOrWatch } from '../../components/ConnectOrWatch'
-import { WATCH_ACCOUNT_ANCHOR } from '../Settings/Settings'
+import { ConnectOrWatch } from '../../components/ConnectCreateOrWatch'
 import { useDisplayLoader } from '../../hooks/useDisplayLoader'
 import { useDisplayError } from '../../hooks/useDisplayError'
 // import CurrentReferendumBanner from '../../components/CurrentReferendumBanner'
@@ -23,14 +18,11 @@ interface HomeProps {
 }
 
 const Home = ({ className }: HomeProps) => {
-  const navigate = useNavigate()
   const [searchParams, setSearchParams] = useSearchParams({
     creationInProgress: 'false'
   })
   const { multiProxyList } = useMultiProxy()
   const [showNewMultisigAlert, setShowNewMultisigAlert] = useState(false)
-  const { isAllowedToConnectToExtension, allowConnectionToExtension } = useAccounts()
-  const { watchedAddresses } = useWatchedAddresses()
   const DisplayError = useDisplayError()
   const DisplayLoader = useDisplayLoader()
 
@@ -54,25 +46,6 @@ const Home = ({ className }: HomeProps) => {
 
   if (DisplayError) {
     return DisplayError
-  }
-
-  if (!isAllowedToConnectToExtension && watchedAddresses.length === 0) {
-    return (
-      <CenterStyled>
-        <h1>Multix is an interface to easily manage complex multisigs.</h1>
-        <p>Connect an extension to interact with Multix or watch an address.</p>
-        <ConnectButtonWrapperStyled>
-          <Button onClick={allowConnectionToExtension}>Connect Wallet</Button>
-          or
-          <Button
-            onClick={() => navigate(`/settings${WATCH_ACCOUNT_ANCHOR}`)}
-            data-cy="button-watch-address"
-          >
-            Watch an address
-          </Button>
-        </ConnectButtonWrapperStyled>
-      </CenterStyled>
-    )
   }
 
   if (multiProxyList.length === 0) {
@@ -149,20 +122,6 @@ const TransactionsWrapperStyled = styled('div')`
   @media (min-width: ${({ theme }) => theme.breakpoints.values.md}px) {
     margin-left: 1.5rem;
   }
-`
-
-const ConnectButtonWrapperStyled = styled('div')`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  & > button {
-    margin: 0 1rem;
-  }
-`
-
-const CenterStyled = styled(Center)`
-  text-align: center;
 `
 
 export default Home
